Move test to test_p_r.py
